refactor(types): move the Worker outcome record beside the failure taxonomy
WorkerOutcome embeds FailureReason/InfraReason but lived in the Worker
adapter, forcing classify.ts into a type-only import cycle and pr.ts/act.ts
to reach into an adapter for a data type. Move it into types.ts next to the
taxonomy it depends on; classify.ts now imports from types.ts only.

Closes #34
